+++ docs/reference/changelog.rst
@@ -19,6 +19,17 @@ Changelog
 
   For a complete list of commits, check out the `X.Y.Z`_ release on GitHub.
 
+.. _release-2.31.0:
+
+2.31.0 (unreleased)
+-------------------
+
+New features:
+
+- The ``OsRelease`` utility can now return the version codename.
+
+For a complete list of commits, check out the `2.31.0`_ release on GitHub.
+
 .. _release-2.30.0:
 
 2.30.0 (2026-03-18)
@@ -1628,6 +1639,7 @@ For a complete list of commits, check out the `2.0.0`_ release on GitHub.
 .. _craft-cli issue #172: https://github.com/canonical/craft-cli/issues/172
 .. _Poetry: https://python-poetry.org
 
+.. _2.31.0: https://github.com/canonical/craft-parts/releases/tag/2.31.0
 .. _2.30.0: https://github.com/canonical/craft-parts/releases/tag/2.30.0
 .. _2.29.0: https://github.com/canonical/craft-parts/releases/tag/2.29.0
 .. _2.28.1: https://github.com/canonical/craft-parts/releases/tag/2.28.1